Split MPNowPlayingInfoCenterPatches
