+++ AudioBar/MPNowPlayingInfoCenter+ElapsedPlaybackTime.swift
@@ -2,8 +2,6 @@ import MediaPlayer
 
 extension MPNowPlayingInfoCenter {
 
-    // MARK: Set elapsed playback time
-
     func setElapsedPlaybackTime(_ elapsedPlaybackTime: TimeInterval) {
         guard needsSetElapsedPlaybackTime(elapsedPlaybackTime) else { return }
         forceSetElapsedPlaybackTime(elapsedPlaybackTime)
@@ -15,21 +13,6 @@ extension MPNowPlayingInfoCenter {
 
     private func forceSetElapsedPlaybackTime(_ elapsedPlaybackTime: TimeInterval) {
         nowPlayingInfo![MPNowPlayingInfoPropertyElapsedPlaybackTime] = elapsedPlaybackTime
-    }
-
-    // MARK: Set playback duration
-
-    func setPlaybackDuration(_ playbackDuration: TimeInterval) {
-        guard needsSetPlaybackDuration(playbackDuration) else { return }
-        forceSetPlaybackDuration(playbackDuration)
-    }
-
-    private func needsSetPlaybackDuration(_ playbackDuration: TimeInterval) -> Bool {
-        return nowPlayingInfo?[MPMediaItemPropertyPlaybackDuration] as? TimeInterval != playbackDuration
-    }
-
-    private func forceSetPlaybackDuration(_ playbackDuration: TimeInterval) {
-        nowPlayingInfo![MPMediaItemPropertyPlaybackDuration] = playbackDuration
     }
 
 }
